Fixed trial 26+ bug
Fixed the problem with Whack-A-T where it wouldn't complete a full day
of trials.

Ended up being an issue with the config generator, a couple of 10s had
to be changed to 20 (because of the increased trial numbers).

I have not hard coded this, so further changes in conditions should not
cause this error anymore.