Fun. for config fix working and implemented
The users individual configuration is now fully read by both the
decision making games.

The order of the fractals is still set by global config. But probability
of being 0 is set by probability taken from users config.

There was a big problem that needed fixing, where the indexing which
each game used to pick fractals, was the same index used to determine
location on the screen (which was shuffled after every stage of the
game). This is now worked around.

The actual probabilities are not reflective of those given, due to the
limit in trial numbers (i.e. there is no way of having 0.7 of 4 full
fractals).